Pass in url as a prop instead of using separate components; rename constants file
